Refactor: Enhance Top Picks screen navigation and interactions
