+++ presentation/src/main/java/com/london/presentation/feature/details/actordetails/topmoviespicks/TopMoviesPicksContract.kt
@@ -2,6 +2,7 @@ package com.london.presentation.feature.details.actordetails.topmoviespicks
 
 interface TopMoviesPicksContract {
     fun onSaveMovie(movieId: Int)
+    fun onMovieClicked(movieId: Int)
     fun onBack()
     fun onRetry()
 }
