+++ presentation/src/main/java/com/london/presentation/feature/details/actordetails/toptvshowspicks/TopTvShowsPicksContract.kt
@@ -1,7 +1,7 @@
 package com.london.presentation.feature.details.actordetails.toptvshowspicks
 
 interface TopTvShowsPicksContract {
-    fun onSaveMovie(movieId: Int)
+    fun onSaveTvShow(tvShowId: Int)
     fun onTvShowClicked(tvShowId: Int)
     fun onBack()
     fun onRetry()
